backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.8.0
djangorestframework-simplejwt==5.5.1
python-decouple==3.8
Pillow==11.3.0
drf-spectacular==0.27.2

# AI and ML libraries
openai==1.54.4
transformers==4.46.3
torch==2.5.1
sentence-transformers==3.3.1
faiss-cpu==1.9.0
numpy==1.26.4
scipy==1.14.1

# Additional utilities
requests==2.32.3
python-dotenv==1.0.1
psycopg2-binary==2.9.10